requirements.txt
python-telegram-bot==20.7
requests
fuzzywuzzy
python-Levenshtein
fastapi
uvicorn
